refactor: split session/mod.rs logic into focused sub-modules
Extract 8 sub-modules from the 1,072-line session/mod.rs:
- key_handlers.rs (280L): handle_key, handle_idle, handle_composing, handle_backspace
- composing.rs (128L): handle_composing_text, append_and_convert, drain/flush
- candidate_gen.rs (84L): update_candidates, deferred response, receive_candidates
- commit.rs (105L): commit_composed, commit_current_state, record_history, reset_state
- submode.rs (87L): toggle_submode with crystallization and boundary space
- auto_commit.rs (170L): try_auto_commit with stability tracking
- ghost.rs (45L): accept/receive ghost text
- response.rs (58L): marked text and candidate response builders

mod.rs now contains only the InputSession struct definition and
core public API methods (136 lines). All sub-modules use
pub(super) visibility for internal methods.
